libpayload/libc/hexdump.c: Change type of `length` argument to `size_t`
Representing a (non-negative) length with a signed integer is not
optimal, so change its type to `size_t`.

The same change was applied to coreboot’s implementation of `hexdump()`
in commit “lib/hexdump: Take const void * and size_t as arguments”
(3dd0e72d) [1].

[1] http://review.coreboot.org/4575

+++ payloads/libpayload/libc/hexdump.c
@@ -29,7 +29,7 @@
 
 #include <libpayload.h>
 
-void hexdump(void *memory, int length)
+void hexdump(void *memory, size_t length)
 {
 	int i;
 	uint8_t *m;
